Bug fixed and add two more sensors

+++ Core/Inc/motor.h
@@ -7,12 +7,12 @@ extern double kp;
 extern double ki;
 extern double kd;
 extern int previous_value;
-extern int integral;
-extern int left_speed;
-extern int right_speed;
-extern int derivative;
+extern double integral;
+extern double left_speed;
+extern double right_speed;
+extern double derivative;
 
-void motor_control(size_t left_motor_speed, size_t right_motor_speed);
+void motor_control(double left_motor_speed, double right_motor_speed);
 void PID_Handle();
 
 #endif
